refactor: split ai feature domain from feature store service
